VULCAN-79 Move `Consumer` classes into `consumer` package
