+++ modules/apps/foundation/vulcan/vulcan-api/bnd.bnd
@@ -3,6 +3,7 @@ Bundle-SymbolicName: com.liferay.vulcan.api
 Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.vulcan.alias,\
+	com.liferay.vulcan.consumer,\
 	com.liferay.vulcan.converter,\
 	com.liferay.vulcan.endpoint,\
 	com.liferay.vulcan.error,\
