+++ modules/apps/foundation/vulcan/vulcan-api/src/main/java/com/liferay/vulcan/consumer/HexaConsumer.java
@@ -12,25 +12,25 @@
  * details.
  */
 
-package com.liferay.vulcan.function;
+package com.liferay.vulcan.consumer;
 
 import java.util.Objects;
 
 /**
- * Represents an operation that accepts nine input arguments and returns no
- * result. This is the nine-arity specialization of {@link
+ * Represents an operation that accepts six input arguments and returns no
+ * result. This is the six-arity specialization of {@link
  * java.util.function.Consumer}. Unlike most other functional interfaces, {@code
  * TriConsumer} is expected to operate via side-effects.
  *
  * <p>This is a <a href="package-summary.html">functional interface</a>
  * whose functional method is {@link #accept(Object, Object, Object, Object,
- * Object, Object, Object, Object, Object)}.
+ * Object, Object)}.
  *
  * @author Alejandro Hernández
  * @see    java.util.function.Consumer
  */
 @FunctionalInterface
-public interface EnneaConsumer<A, B, C, D, E, F, G, H, I> {
+public interface HexaConsumer<A, B, C, D, E, F> {
 
 	/**
 	 * Performs this operation on the given arguments.
@@ -41,32 +41,29 @@ public interface EnneaConsumer<A, B, C, D, E, F, G, H, I> {
 	 * @param d the fourth function argument
 	 * @param e the fifth function argument
 	 * @param f the sixth function argument
-	 * @param g the seventh function argument
-	 * @param h the eighth function argument
-	 * @param i the ninth function argument
 	 */
-	public void accept(A a, B b, C c, D d, E e, F f, G g, H h, I i);
+	public void accept(A a, B b, C c, D d, E e, F f);
 
 	/**
-	 * Returns a composed {@code EnneaConsumer} that performs, in sequence, this
+	 * Returns a composed {@code HexaConsumer} that performs, in sequence, this
 	 * operation followed by the {@code after} operation. If performing either
 	 * operation throws an exception, it is relayed to the caller of the
 	 * composed operation. If performing this operation throws an exception, the
 	 * {@code after} operation will not be performed.
 	 *
 	 * @param  after the operation to perform after this operation
-	 * @return a composed {@code EnneaConsumer} that performs in sequence this
+	 * @return a composed {@code HexaConsumer} that performs in sequence this
 	 *         operation followed by the {@code after} operation
 	 */
-	public default EnneaConsumer<A, B, C, D, E, F, G, H, I> andThen(
-		EnneaConsumer<? super A, ? super B, ? super C, ? super D, ? super E,
-			? super F, ? super G, ? super H, ? super I> after) {
+	public default HexaConsumer<A, B, C, D, E, F> andThen(
+		HexaConsumer<? super A, ? super B, ? super C, ? super D, ? super E,
+			? super F> after) {
 
 		Objects.requireNonNull(after);
 
-		return (a, b, c, d, e, f, g, h, i) -> {
-			accept(a, b, c, d, e, f, g, h, i);
-			after.accept(a, b, c, d, e, f, g, h, i);
+		return (a, b, c, d, e, f) -> {
+			accept(a, b, c, d, e, f);
+			after.accept(a, b, c, d, e, f);
 		};
 	}
