+++ modules/apps/foundation/vulcan/vulcan-api/src/main/java/com/liferay/vulcan/consumer/TriConsumer.java
@@ -12,57 +12,53 @@
  * details.
  */
 
-package com.liferay.vulcan.function;
+package com.liferay.vulcan.consumer;
 
 import java.util.Objects;
 
 /**
- * Represents an operation that accepts five input arguments and returns no
- * result. This is the five-arity specialization of {@link
+ * Represents an operation that accepts three input arguments and returns no
+ * result. This is the three-arity specialization of {@link
  * java.util.function.Consumer}. Unlike most other functional interfaces, {@code
  * TriConsumer} is expected to operate via side-effects.
  *
  * <p>This is a <a href="package-summary.html">functional interface</a>
- * whose functional method is {@link #accept(Object, Object, Object, Object,
- * Object)}.
+ * whose functional method is {@link #accept(Object, Object, Object)}.
  *
  * @author Alejandro Hernández
  * @see    java.util.function.Consumer
  */
 @FunctionalInterface
-public interface PentaConsumer<A, B, C, D, E> {
+public interface TriConsumer<A, B, C> {
 
 	/**
 	 * Performs this operation on the given arguments.
 	 *
-	 * @param a the first function argument
-	 * @param b the second function argument
-	 * @param c the third function argument
-	 * @param d the fourth function argument
-	 * @param e the fifth function argument
+	 * @param a the first input argument.
+	 * @param b the second input argument.
+	 * @param c the third input argument.
 	 */
-	public void accept(A a, B b, C c, D d, E e);
+	public void accept(A a, B b, C c);
 
 	/**
-	 * Returns a composed {@code PentaConsumer} that performs, in sequence, this
+	 * Returns a composed {@code TriConsumer} that performs, in sequence, this
 	 * operation followed by the {@code after} operation. If performing either
 	 * operation throws an exception, it is relayed to the caller of the
 	 * composed operation. If performing this operation throws an exception, the
 	 * {@code after} operation will not be performed.
 	 *
 	 * @param  after the operation to perform after this operation
-	 * @return a composed {@code PentaConsumer} that performs in sequence this
+	 * @return a composed {@code TriConsumer} that performs in sequence this
 	 *         operation followed by the {@code after} operation
 	 */
-	public default PentaConsumer<A, B, C, D, E> andThen(
-		PentaConsumer<? super A, ? super B, ? super C, ? super D, ? super E>
-			after) {
+	public default TriConsumer<A, B, C> andThen(
+		TriConsumer<? super A, ? super B, ? super C> after) {
 
 		Objects.requireNonNull(after);
 
-		return (a, b, c, d, e) -> {
-			accept(a, b, c, d, e);
-			after.accept(a, b, c, d, e);
+		return (a, b, c) -> {
+			accept(a, b, c);
+			after.accept(a, b, c);
 		};
 	}
